worker: split osbuild job implementation into separate file
No code was changed.
